add initial version of sider component
