+++ app/src/main/java/com/bck/handshake/Account.kt
@@ -17,54 +17,64 @@ import com.bck.handshake.data.sampleRecords
 //navController.navigate(NewBetScreen().route)
 
 @Composable
-
-
-
 fun AccountScreen(
     onNewBetClicked: () -> Unit,
     modifier: Modifier = Modifier
 ) {
-
     Surface(
         color = MaterialTheme.colorScheme.background,
         modifier = Modifier.fillMaxSize()
     ) {
-        // Create a Column layout to hold the user image and the subheader.
         Column(
             modifier = Modifier.fillMaxSize(),
-            horizontalAlignment = Alignment.CenterHorizontally
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.SpaceBetween
         ) {
+            Column(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                // create an Icon that used the Person icon from the Material Icons library.
+                Icon(
+                    imageVector = Icons.Filled.Person,
+                    contentDescription = "User image",
+                    modifier = Modifier
+                        .padding(top = 16.dp)
+                        .size(86.dp)
+                        .clip(CircleShape)
+                        .border(2.dp, Color.Black, CircleShape)
+                )
+                // Add the subheader.
+                Text(
+                    text = sampleRecords.formattedRecords,
+                    modifier = Modifier.padding(16.dp),
+                    style = MaterialTheme.typography.labelMedium
+                )
+                Spacer(modifier = Modifier.height(16.dp))
+                Button(
+                    onClick = onNewBetClicked,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(48.dp)
+                        .padding(horizontal = 24.dp)
+                ) {
+                    Text(text = "New Bet")
+                }
+            }
 
-
-            // create an Icon that used the Person icon from the Material Icons library.
-            Icon(
-                imageVector = Icons.Filled.Person,
-                contentDescription = "User image",
-                modifier = Modifier
-                    .padding(top = 16.dp)
-                    .size(86.dp)
-                    .clip(CircleShape)
-                    .border(2.dp, Color.Black, CircleShape)
-            )
-            // Add the subheader.
-            Text(
-                text = sampleRecords.formattedRecords,
-                modifier = Modifier.padding(16.dp),
-                style = MaterialTheme.typography.labelMedium
-            )
-            Spacer(modifier = Modifier.height(16.dp))
-            Button(
-                onClick = {
-                },
+            // Add HandshakeSlider at the bottom
+            Column(
                 modifier = Modifier
                     .fillMaxWidth()
-                    .height(48.dp)
-                    .padding(horizontal = 24.dp)
+                    .padding(bottom = 26.dp),
+                horizontalAlignment = Alignment.CenterHorizontally
             ) {
-                Text(text = "New Bet")
+                HandshakeSlider(
+                    onConfirmed = onNewBetClicked,
+                    modifier = Modifier.fillMaxWidth()
+                )
             }
         }
-
     }
 }
